Add nav cards and goldstar integration, rename component
