+++ src/ui/design-system/src/index.ts
@@ -3,7 +3,7 @@ export { Heading } from './lib/Components/Heading'
 export type { HeadingProps } from './lib/Components/Heading'
 // export { InternalSidebar } from './lib/Components/InternalSidebar'
 export { LandingHeader } from './lib/Components/LandingHeader'
-export { LandingHeaderHome } from './lib/Components/LandingHeaderHome'
+export { LandingHeaderHome } from './lib/Components/LandingNav'
 export { LinkCard2Column } from './lib/Components/LinkCard2Column'
 export type { LinkCard2ColumnProps } from './lib/Components/LinkCard2Column'
 export { default as NetworkCard } from './lib/Components/NetworkCard'
